Add formatting for the book

+++ NET.W.2017.Buiko.08/NET.W.2017.Buiko.08.01/Logic/Storage/Implementation/BinaryFileBookStorage.cs
@@ -78,7 +78,7 @@ namespace Logic.Storage.Implementation
             var publishingHouse = binaryReader.ReadString();
             var publicationYear = binaryReader.ReadString();
             var pageNumber = binaryReader.ReadInt32();
-            var price = binaryReader.ReadInt32();
+            var price = binaryReader.ReadDecimal();
 
             return new Book(isbn, author, name, publishingHouse, publicationYear, pageNumber, price);
         }
